+++ Nicuga/Hostile.cpp
@@ -142,8 +142,8 @@ void HostileInstance::run(long dt)
 		}
 	}
 
-	float posx = this->initialPosition.getX() + FastMath::fastCos((int)direction + 90) * this->speed * (float)this->age * 1/1000.0f;
-	float posy = this->initialPosition.getY() + FastMath::fastSin((int)direction + 90) * this->speed * (float)this->age * 1/1000.0f;
+	short posx = this->initialPosition.getX() + FastMath::fastCos((int)direction + 90) * this->speed * (float)this->age * 1/1000.0f;
+	short posy = this->initialPosition.getY() + FastMath::fastSin((int)direction + 90) * this->speed * (float)this->age * 1/1000.0f;
 	//printf("HOSTILE INSTANCE POS(%f,%f)\n",direction,(float)this->age);
 	this->position.set(posx,posy);
 	for (unsigned int i = 0;i < this->barrageList.size();i++)
